requirements-all.txt
decorator==4.0.6
hy==0.12.1
docopts==0.6.1
PyHamcrest==1.9.0
mockito-without-hardcoded-distribute-version==0.5.4
behave==1.2.5
nose==1.3.7
nose-cov==1.6
python-coveralls==2.6.0
sphinx==1.3.5
hymn==0.6
hypothesis==3.6.1
libarchimedes==0.4.1
toolz==0.7.4
